Convert to C++

+++ src/audio.h
@@ -13,7 +13,7 @@ int audioSoundDecoderReadHandler(int fileHandle, void* buf, unsigned int size);
 int audioOpen(const char* fname, int mode, ...);
 int audioClose(int fileHandle);
 int audioRead(int fileHandle, void* buffer, unsigned int size);
-int audioSeek(int fileHandle, long offset, int origin);
+long audioSeek(int fileHandle, long offset, int origin);
 long audioGetSize(int fileHandle);
 long audioTell(int fileHandle);
 int audioWrite(int handle, const void* buf, unsigned int size);
